# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "URL Shortener API",
    "version": "1.0.0",
    "description": "Create short links, redirect them, inspect click logs, export CSV, and delete links."
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": { "description": "Service is healthy" }
        }
      }
    },
    "/links": {
      "post": {
        "summary": "Create a short link",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "target_url": { "type": "string" },
                  "code": { "type": "string" },
                  "expires_at": { "type": "string", "format": "date-time" }
                },
                "required": ["target_url"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Link created" },
          "400": { "description": "Validation failed" },
          "409": { "description": "Code already taken" }
        }
      }
    },
    "/{code}": {
      "get": {
        "summary": "Redirect a short code",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "302": { "description": "Redirect to target" },
          "404": { "description": "Link not found" },
          "410": { "description": "Link expired" }
        }
      }
    },
    "/links/{code}": {
      "get": {
        "summary": "Fetch link metadata",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Link metadata" },
          "404": { "description": "Link not found" }
        }
      },
      "delete": {
        "summary": "Delete a link",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Link deleted" },
          "404": { "description": "Link not found" }
        }
      }
    },
    "/links/{code}/clicks": {
      "get": {
        "summary": "Fetch paginated click log",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "after", "in": "query", "schema": { "type": "string" } },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": { "description": "Paginated click log" },
          "404": { "description": "Link not found" }
        }
      }
    },
    "/links/{code}/clicks.csv": {
      "get": {
        "summary": "Export click log as CSV",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "CSV stream" } }
      }
    }
  }
}